# Cache compiler lookup per package