@sethrj 2021-05-28
@alalazo weird error from mypy:
```
lib/spack/spack/package.py:1272: error: Decorated property not supported
```
is that an error in the package, or an internal limitation in mypy?