@alalazo 2021-05-28
@sethrj See https://github.com/spack/spack/pull/23988#discussion_r641692092